fix crashes in test::ReadSDO
RTT calls cleanupHook twice in the event of an exception. Once during
the running->exception transition and once during exception->pre_operational

I always assumed that cleanupHook would be called only once per
call to configureHook *and that's it*. But it's not the case.

In test::ReadSDO's configureHook, this led to a double-free